requirements.txt
llama-cpp-python>=0.2.0
tqdm>=4.66.0
numpy>=1.24.0
pyttsx3>=2.90
opencv-python>=4.8.0
Pillow>=10.0.0 